// src/app/shared/store/task/task.type.ts
export enum ProjectTaskType {
    LIST = '[PROJECT TASK API] List',
    LIST_SUCCESS = '[PROJECT TASK API] List Success',
    LIST_CANCEL = '[PROJECT TASK API] List Cancel',

    DETAILS = '[PROJECT TASK API] Details',
    DETAILS_SUCCESS = '[PROJECT TASK API] Details Success',
    DETAILS_CANCEL = '[PROJECT TASK API] Details Cancel',

    CREATE = '[PROJECT TASK API] Create',
    CREATE_SUCCESS = '[PROJECT TASK API] Create Success',
    CREATE_CANCEL = '[PROJECT TASK API] Create Cancel',

    UPDATE = '[PROJECT TASK API] Update',
    UPDATE_SUCCESS = '[PROJECT TASK API] Update Success',
    UPDATE_CANCEL = '[PROJECT TASK API] Update Cancel',

    DELETE = '[PROJECT TASK API] Delete',
    DELETE_SUCCESS = '[PROJECT TASK API] Delete Success',
    DELETE_CANCEL = '[PROJECT TASK API] Delete Cancel'
}